Added / removed Django management commands.
